requirements.txt
blinker==1.9.0
cachelib==0.13.0
click==8.1.7
colorama==0.4.6
Flask==3.1.0
Flask-Session==0.8.0
itsdangerous==2.2.0
Jinja2==3.1.4
Markdown==3.7
MarkupSafe==3.0.2
msgspec==0.18.6
psycopg2-binary==2.9.10
Werkzeug==3.1.3
